abstracted the theme system further
A base interface, ChatHandler, was created to further abstract the
functionality of a Theme. Hopefully this will allow plugins to add
ChatHandlers that don't have to follow the header-body-footer format of
a Theme.

Some other things:
- ThemeElement has been removed. The Theme class (now abstract) will
require you to override getHeaderFormat(), getBodyFormat(), and
getFooterFormat() instead.
- VanillaTheme has been moved out of ThemeManager and is now its own
class.
- Since Theme now requires a HyperChat instance, ThemeManager now also
requires a HyperChat instance.
- User has been modified to use ChatHandlers instead of Themes.